+++ sdk/core/azure-core/src/http/curl/curl.cpp
@@ -1077,7 +1077,7 @@ inline std::string GetConnectionKey(std::string const& host, CurlTransportOption
   {
     key.append("0");
   }
-  if (options.SSLOptions.NoRevoke)
+  if (!options.SSLOptions.EnableCertificateRevocationListCheck)
   {
     key.append("1");
   }
@@ -1202,7 +1202,7 @@ std::unique_ptr<CurlNetworkConnection> CurlConnectionPool::GetCurlConnection(
   }
 
   long sslOption = 0;
-  if (options.SSLOptions.NoRevoke)
+  if (!options.SSLOptions.EnableCertificateRevocationListCheck)
   {
     sslOption |= CURLSSLOPT_NO_REVOKE;
   }