+++ sdk/core/azure-core/test/ut/uuid.cpp
@@ -11,7 +11,7 @@ using namespace Azure::Core;
 TEST(Uuid, Basic)
 {
   auto uuid = Uuid::CreateUuid();
-  EXPECT_TRUE(uuid.GetUuidString().length() == 36);
+  EXPECT_TRUE(uuid.ToString().length() == 36);
 }
 
 TEST(Uuid, Randomness)
@@ -20,7 +20,7 @@ TEST(Uuid, Randomness)
   std::set<std::string> uuids;
   for (int i = 0; i < size; i++)
   {
-    auto ret = uuids.insert(Uuid::CreateUuid().GetUuidString());
+    auto ret = uuids.insert(Uuid::CreateUuid().ToString());
     // If the value already exists in the set then the insert will fail
     // ret.second == false means the insert failed.
     EXPECT_TRUE(ret.second);
@@ -30,7 +30,7 @@ TEST(Uuid, Randomness)
 
 TEST(Uuid, separatorPosition)
 {
-  auto uuidKey = Uuid::CreateUuid().GetUuidString();
+  auto uuidKey = Uuid::CreateUuid().ToString();
   // validate expected format '8-4-4-4-12'
   EXPECT_PRED5(
       [](std::string const& uuidKey, char pos1, char pos2, char pos3, char pos4) {
@@ -45,7 +45,7 @@ TEST(Uuid, separatorPosition)
 
 TEST(Uuid, validChars)
 {
-  auto uuidKey = Uuid::CreateUuid().GetUuidString();
+  auto uuidKey = Uuid::CreateUuid().ToString();
   // validate valid chars and separators count
   EXPECT_PRED2(
       [](std::string const& uuidKey, int expectedSeparators) {