+++ sdk/storage/azure-storage-common/src/storage_per_retry_policy.cpp
@@ -25,7 +25,7 @@ namespace Azure { namespace Storage { namespace Details {
       request.AddHeader(
           HttpHeaderXMsDate,
           Core::DateTime(std::chrono::system_clock::now())
-              .GetString(Azure::Core::DateTime::DateFormat::Rfc1123));
+              .ToString(Azure::Core::DateTime::DateFormat::Rfc1123));
     }
 
     return nextHttpPolicy.Send(ctx, request);