+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_client.cpp
@@ -468,7 +468,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     if (options.ExpiresOn.HasValue())
     {
       protocolLayerOptions.ExpiryTime
-          = options.ExpiresOn.GetValue().GetString(Azure::Core::DateTime::DateFormat::Rfc1123);
+          = options.ExpiresOn.GetValue().ToString(Azure::Core::DateTime::DateFormat::Rfc1123);
     }
     else if (options.TimeToExpire.HasValue())
     {