+++ sdk/core/azure-core/CHANGELOG.md
@@ -7,6 +7,9 @@
 - Removed `Azure::Core::Http::HttpPipeline` by making it internal, used only within the SDK.
 - Split `Azure::Core::RequestConditions` into `Azure::Core::MatchConditions` and `Azure::Core::ModifiedConditions`.
 - Removed `TransportKind` enum from `Azure::Core::Http`.
+- Renamed `NoRevoke` to `EnableCertificateRevocationListCheck` for `Azure::Core::Http::CurlTransportSSLOptions`.
+- Renamed `GetString()` to `ToString()` in `Azure::Core::DateTime`.
+- Renamed `GetUuidString()` tp `ToString()` in `Azure::Core::Uuid`.
 
 ## 1.0.0-beta.6 (2021-02-09)